Moving from 32-bit to 64-bit return value of the file close method. Last modification which will allow to write big file in alimdc and mStreamRecorder

+++ RAW/mdc.cxx
@@ -58,7 +58,7 @@ long long alimdcGetTotalFileSize(void* alimdc)
   return ((AliMDC*)alimdc)->GetTotalSize();
 }
 
-int alimdcClose(void* alimdc)
+long long alimdcClose(void* alimdc)
 {
 // close the raw DB
